More fluid scrolling, continued.
Reverse and push further the idea of offscreen preparation as done in commit f30312bbe553a243e8be99fa5cfaba91f381ff44.
We add and remove cells in a view hierarchy which from now has a height of three times the height of the visible rectangle; so we are able to quickly respond to `pageDown` and `pageUp` requests.
Fix https://github.com/ViennaRSS/vienna-rss/issues/229 ; should probably fix issue #20 too.